requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --annotation-style=line requirements.in -o requirements.txt --no-strip-extras
agate==1.7.1              # via dbt-adapters, dbt-common, dbt-core, dbt-snowflake
annotated-types==0.7.0    # via pydantic
appdirs==1.4.4            # via sqlfluff
asn1crypto==1.5.1         # via snowflake-connector-python
attrs==24.3.0             # via jsonschema, referencing
babel==2.16.0             # via agate
backports-tarfile==1.2.0  # via jaraco-context
certifi==2024.12.14       # via requests, snowflake-connector-python
cffi==1.17.1              # via cryptography, snowflake-connector-python
chardet==5.2.0            # via diff-cover, sqlfluff
charset-normalizer==3.4.0  # via requests, snowflake-connector-python
click==8.1.7              # via dbt-core, dbt-semantic-interfaces, sqlfluff
colorama==0.4.6           # via dbt-common, sqlfluff
cryptography==44.0.0      # via pyopenssl, snowflake-connector-python
daff==1.3.46              # via dbt-core
dbt-adapters==1.14.0      # via dbt-core, dbt-snowflake
dbt-common==1.14.0        # via dbt-adapters, dbt-core, dbt-snowflake
dbt-core==1.9.2           # via dbt-snowflake, -r requirements.in
dbt-extractor==0.5.1      # via dbt-core, -r requirements.in
dbt-semantic-interfaces==0.7.4  # via dbt-core, -r requirements.in
dbt-snowflake==1.9.1      # via -r requirements.in
deepdiff==7.0.1           # via dbt-common
diff-cover==9.2.1         # via sqlfluff
duckdb==1.2.1             # via -r requirements.in
filelock==3.16.1          # via snowflake-connector-python
idna==3.10                # via requests, snowflake-connector-python
importlib-metadata==6.11.0  # via dbt-semantic-interfaces, keyring
iniconfig==2.0.0          # via pytest
isodate==0.6.1            # via agate, dbt-common
jaraco-classes==3.4.0     # via keyring
jaraco-context==6.0.1     # via keyring
jaraco-functools==4.1.0   # via keyring
jinja2==3.1.5             # via dbt-common, dbt-core, dbt-semantic-interfaces, diff-cover, sqlfluff, -r requirements.in
jsonschema==4.23.0        # via dbt-common, dbt-semantic-interfaces
jsonschema-specifications==2024.10.1  # via jsonschema
keyring==25.5.0           # via snowflake-connector-python
leather==0.4.0            # via agate
markupsafe==3.0.2         # via jinja2
mashumaro[msgpack]==3.14  # via dbt-adapters, dbt-common, dbt-core
more-itertools==10.5.0    # via dbt-semantic-interfaces, jaraco-classes, jaraco-functools
msgpack==1.1.0            # via mashumaro
networkx==3.4.2           # via dbt-core
numpy==1.26.4             # via pandas, -r requirements.in
ordered-set==4.1.0        # via deepdiff
packaging==24.2           # via dbt-core, pytest, snowflake-connector-python
pandas==2.0.2             # via -r requirements.in
parsedatetime==2.6        # via agate
pathspec==0.11.2          # via dbt-common, dbt-core, sqlfluff
platformdirs==4.3.6       # via snowflake-connector-python
pluggy==1.5.0             # via diff-cover, pytest
protobuf==5.29.3          # via dbt-adapters, dbt-common, dbt-core
pycparser==2.22           # via cffi
pydantic==2.10.4          # via dbt-semantic-interfaces
pydantic-core==2.27.2     # via pydantic
pygments==2.18.0          # via diff-cover
pyjwt==2.10.1             # via snowflake-connector-python
pyopenssl==24.3.0         # via snowflake-connector-python
pytest==8.3.4             # via sqlfluff
python-dateutil==2.9.0.post0  # via dbt-common, dbt-semantic-interfaces, pandas
python-slugify==8.0.4     # via agate
pytimeparse==1.1.8        # via agate
pytz==2024.2              # via dbt-adapters, dbt-core, pandas, snowflake-connector-python
pyyaml==6.0.2             # via dbt-core, dbt-semantic-interfaces, sqlfluff
referencing==0.35.1       # via jsonschema, jsonschema-specifications
regex==2024.11.6          # via sqlfluff
requests==2.32.3          # via dbt-common, dbt-core, snowflake-connector-python, snowplow-tracker
rpds-py==0.22.3           # via jsonschema, referencing
ruamel-yaml==0.18.14      # via -r requirements.in
ruamel-yaml-clib==0.2.12  # via ruamel-yaml
six==1.17.0               # via isodate, python-dateutil
snowflake-connector-python[secure-local-storage]==3.12.4  # via dbt-snowflake, -r requirements.in
snowplow-tracker==1.0.4   # via dbt-core
sortedcontainers==2.4.0   # via snowflake-connector-python
sqlfluff==3.1.0           # via -r requirements.in
sqlparse==0.5.3           # via dbt-core
tblib==3.0.0              # via sqlfluff
text-unidecode==1.3       # via python-slugify
tomlkit==0.13.2           # via snowflake-connector-python
tqdm==4.67.1              # via sqlfluff
types-requests==2.31.0.6  # via snowplow-tracker
types-urllib3==1.26.25.14  # via types-requests
typing-extensions==4.12.2  # via dbt-adapters, dbt-common, dbt-core, dbt-semantic-interfaces, mashumaro, pydantic, pydantic-core, snowflake-connector-python, snowplow-tracker
tzdata==2024.2            # via pandas
urllib3==1.26.20          # via requests
zipp==3.21.0              # via importlib-metadata
